requirements.txt
aiofiles==0.6.0
certifi==2020.6.20
chardet==3.0.4
greenlet==0.4.17
h11==0.9.0
httpcore==0.11.1
httptools==0.1.1
httpx==0.15.4
idna==2.10
multidict==5.0.0
PyJWT==1.7.1
rfc3986==1.4.0
sanic==20.9.1
sniffio==1.2.0
SQLAlchemy==1.4.0b1
urllib3==1.25.11
websockets==8.1
